feat: restructure API server by consolidating app and server logic into app.ts and removing server.ts
