--- src/server.ts
@@ -1,7 +1,0 @@
-import app from "./index";
-
-const server = app.listen(3000, () => {
-  console.log("Server is running on port 3000");
-});
-
-export default server;
